Refactor client code into its own file
